pyproject.toml
[project]
name = "docling-langextract-testing"
version = "0.1.0"
description = "Testing docling + langextract combination for paralegal date extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=0.1.0",
    "extract-msg>=0.55.0",
    "langextract",
    "numpy<2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "regex>=2025.9.18",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "textstat>=0.7.10",
    "torch>=2.0.0,<2.3.0",
]

[tool.uv.sources]
langextract = { git = "https://github.com/google/langextract.git" }
